pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biometric-voice"
version = "0.1.0"
description = "Biometric voice recognition system using SpeechBrain"
requires-python = ">=3.9"
dependencies = [
    "speechbrain>=1.0.0",
    "torch>=2.0.0,<2.5.0",
    "torchaudio>=2.0.0,<2.5.0",
    "numpy",
    "soundfile",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "huggingface_hub<0.26.0",
    "psycopg2-binary",
    "pgvector",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
record = ["sounddevice"]

[project.scripts]
biometric-voice = "biometric_voice.cli:main"
biometric-voice-api = "biometric_voice.api:main"

[tool.setuptools.packages.find]
include = ["biometric_voice*"]
